Jonathan's changes... CUDA Error: Invalid device function
