+++ src/serac/physics/utilities/functional/tests/CMakeLists.txt
@@ -62,7 +62,7 @@ if(ENABLE_CUDA)
 
     set(functional_tests_cuda
         tensor_unit_tests_cuda.cu
-    #    functional_comparisons_cuda.cu
+        functional_comparisons_cuda.cu
     )
         
     foreach(filename ${functional_tests_cuda})
